backend/requirements.txt
fastapi>=0.95.2
uvicorn>=0.22.0
python-multipart>=0.0.6
pytesseract>=0.3.10
pillow>=9.5.0 --only-binary :all:
pdf2image>=1.16.3
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
sqlalchemy>=2.0.19
pydantic>=1.10.12,<2.0.0
python-dotenv>=1.0.0
opencv-python-headless>=4.8.0
numpy>=1.24.0
aiofiles>=23.1.0
openai>=1.0.0
PyMuPDF>=1.22.3
scikit-image>=0.22.0 